# tests: re-enable blockchain GeneralStateTests (state tests converted to blockchain test format).  Disable VMTests VMPerformance suite.