These will certainly be useful for stateless witness construction/verification test coverage.  After disabling the `VMTests/vmPerformance` suite, the test suite runtime only takes 110 seconds.

From my machine:

Blockchain tests without vmPerformance - 110.071s
Blockchain tests with vmPerformance - 594.565s
Blockchain tests without GeneralStatetests (master) - 32.120s

State tests (master) - 59.723s