@karalabe 2024-06-20
This should just be popped on top of the stateless PR